requirements.txt
streamlit==1.24.1
pandas==2.0.3
requests==2.31.0
plotly==5.15.0
pydeck==0.8.0
Pillow==10.0.0
